mv send_email to pipeline.send_email
